Convert PrivacySettingsTable to TypeScript
